Change encryption method of PKCS#12 storage
No longer use a user password to store a PKCS#12 file. Instead,
use the salt and pepper mechanism to safely store the PKCS#12 in
the database. The user should only enter a password at the time
when he wants to download a PKCS#12 file. The new mechanism also
allows us to get the private key and certificate from the database
even if the employee is "lost".